refactor(habitat): split structural execution policies
